+++ src/views/components/l13-diff-intro/l13-diff-intro.component.ts
@@ -24,6 +24,16 @@ const keyboardShortcuts:Shortcut[] = [
 		key: 'Ctrl+F',
 		mac: 'Cmd+F',
 	},
+	{
+		description: 'Delete Selected Files',
+		key: 'Delete',
+		mac: 'Cmd+Backspace',
+	},
+	{
+		description: 'Delete Left Or Right Files',
+		key: 'Alt+Delete',
+		mac: 'Cmd+Alt+Backspace',
+	},
 ];
 
 //	Initialize _________________________________________________________________